added ability to save blockly code to server (#1)
Client code does a POST request containing generated code from blockly
to a php file on the server if the "update robot code" button is
pressed. That file on the server saves the received code to a "code.py"
file in the directory that the php file is running in.